cleanup of interface and class
